+++ modules/react/src/components/index.js
@@ -22,3 +22,4 @@ export {default as EncoderDropdown} from './encoder-dropdown';
 export {default as ResolutionGuide} from './resolution-guide';
 export {ExportVideoModal, ExportVideoPanelContainer} from './export-video';
 export {injectKeplerUI, InjectKeplerUI, WithKeplerUI, KeplerUIContext} from './inject-kepler';
+export {QuickAnimation} from './quick-animation';
